feat: refactor references view stuff into more components
