+++ app/frontend/src/features/profile/view/ProfilePage.tsx
@@ -117,7 +117,7 @@ export default function ProfilePage() {
                   classes={{ root: classes.tabPanel }}
                   value="references"
                 >
-                  <References user={user} />
+                  <References />
                 </TabPanel>
               </TabContext>
             </Card>
